+++ lib/services/js_recorder_bridge.dart
@@ -1,0 +1,2 @@
+export 'js_recorder_bridge_stub.dart'
+    if (dart.library.js_interop) 'js_recorder_bridge_web.dart';